app/Blink/requirements.txt
click
halo
pandas
pillow
seaborn
